Fix whitaker violations in new survivor-kill tests
Move the send-streaming config accessor test into its own module so
send_streaming.rs stays under the 400-line cap, and inline the
declared-total submission into its test function so the expect call
lives inside a test rather than a plain helper.
